The date ticks were being incorrectly calculated.  Alexei gave the data to be added.
Reviewed by Joseph.

+++ trunk/sci2/plugins/normal_plugins/edu.iu.sci2.visualization.temporalbargraph/src/edu/iu/sci2/visualization/temporalbargraph/common/AbstractVisualization.java
@@ -69,7 +69,7 @@ public abstract class AbstractVisualization {
 		int endYear = endDate.toLocalDate().getYear() + 1;
 		
 		int difference = (endYear - startYear);
-		double yearsPerStep = difference / maxDates;
+		double yearsPerStep = (double) difference / (double) maxDates;
 		if (yearsPerStep < 1) {
 			// All the dates will fit
